Next event estimation, diffuse only
It took lots of debugging, but we are here now. Next event estimation implemented, with only diffuse indirect lighting for now. Comparing against the old version, it seems to work as it gets very close to the same amount of total energy received by the sensor compared to the brute-force path tracer.
Also did some refactoring of the primitive functions, and made an abstract primitive type that will automatically call the right functions based on the type.